african star: create covers for the original tokens
The problem with the original tokens is that they're made of a thin
cardboard-like material which makes them susceptible to being
marked. Instead of replaced all the tokens, we can print a plastic
cover and glue the original tokens on the cover.

Also, the original instructions mention there are 30 tokens, but in
reality the game comes with 33 tokens and we're required to remove
three prior to a game.

Signed-off-by: Felipe Balbi <felipe@balbi.sh>